Alright hacked my way into getting this code testable
What we do is take the current repository and copy it into tmp. Then in
tmp we checkout the first commit in this repo 008ecb4 and then use that
to test the remaining two things. Since we now can adjust the cwd of the
git repo we can use it instead of the main file for great success